KAD-512: group engine mode tests into `KadaiEngineModesTest`
